separated tests and stuff
