+++ src/main/java/net/clesperanto/test/YetAnotherPlayground.java
@@ -1,9 +1,10 @@
-package net.clesperanto.clicwrapper;
+package net.clesperanto.test;
 
 import ij.IJ;
 import ij.ImageJ;
 import ij.ImagePlus;
 import net.clesperanto.ClesperantoJ;
+import net.clesperanto.clicwrapper.clesperantojWrapper;
 import net.clesperanto.pyclesperanto;
 
 public class YetAnotherPlayground {
